Cleaning up the code. Don't need to indicate private (class is private by default). Class member functions should be defined in .h
